requirements.txt
numpy
pytorch
torchvision
scikit-learn
torch_dct
cffi
matplotlib
